// src/app/modules/campaign-call-list-validation/models/response/CallValidationResponseModel.ts
export interface CallValidationResponseModel {
    campaignPlayerId: number
    callListNotes: string
    playerId: string
    username: string
    playerStatusName: string
    brand: string
    currency: string
    country: string
    mobileNumber: string
    marketingSource: string
    campaignName: string
    registeredDate: string
    deposited: boolean
    ftdAmount: number
    ftdDate: string
    lastDepositDate: string
    lastDepositAmount: number
    agentName: string
    agentId: number
    taggedBy: string
    taggedDate: string
    primaryGoalReached: boolean
    primaryGoalCount: number
    primaryGoalAmount: number
    primaryGoalAmountInUSD: number
    validIncentivePoints: number
    validIncentiveSourced: string
    validIncentiveSourcedUSD: number
    invalidIncentivePoints: number
    invalidIncentiveSource: string
    invalidIncentiveSourceUSD: number
    incentiveValue: number
    systemValidation: boolean
    agentValidation: boolean
    agentValidationNotes: string
    leaderValidation: boolean
    leaderJustification: string
    leaderValidationNotes: string
    callEvaluationPoint?: number
    callEvaluationNotes: string
    highDepositAmount: number
    lastMessageStatus: string
    lastCallDate: string
    callCount: number
    contactableCallCount: number
    lastContactableCaseDate: string
    totalCallDuration: number
    firstCallDatetime: string
    firstCallStatus: string
    secondCallDatetime: string
    secondCallStatus: string
    lastCallDateTimeAfter2nd: string
    lastCallStatusAfter2nd: string
    firstCallAttemptCount: number
    secondCallAttemptCount: number
    additionalCallAttempt: number
    campaignId: number
    depositAttempts: number
    isAgentValidated: boolean
    isLeaderValidated: boolean
    campaignStatus: number,
    messageStatusAndResponseResult: string
    campaignType: string
    initialDepositAmount: number
    initialDepositDate: string
    totalDepositAmount: number
    totalDepositCount: number
}